+++ satella/exceptions.py
@@ -1,10 +1,13 @@
 class BaseSatellaException(Exception):
     def __init__(self, msg, *args, **kwargs):
-        super().__init__(*((msg,) + args))
+        super().__init__(*(msg, *args))
         self.msg = msg
 
     def __str__(self):
         return '%s(%s)' % (self.__class__.__qualname__, self.args)
+
+    def __repr__(self):
+        return '%s(%s)' % (self.__class__.__qualname__, ', '.join(map(repr, self.args)))
 
 
 class ResourceLocked(BaseSatellaException):